RIC/requirements.txt
asgiref==3.6.0
Babel==2.11.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.6
django-allauth==0.52.0
django-allauth-microsoft==1.0.3
django-auth-adfs==1.11.4
django-bootstrap4==22.3
django-chartjs==2.3.0
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-filter==22.1
django-import-export==3.0.2
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-phonenumber-field==7.0.2
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
msal==1.21.0
numpy==1.24.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.0
pandas==2.0.0
pdfkit==1.0.0
phonenumbers==8.13.5
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
razorpay==1.3.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.3.0
tzdata==2023.3
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
